+++ rtsp-server/xop/AACSource.cpp
@@ -111,8 +111,8 @@ bool AACSource::HandleFrame(const MediaChannelId channel_id,
 	RtpPacket rtp_pkt;
 	rtp_pkt.type = frame.type;
 	rtp_pkt.timestamp = frame.timestamp;
-	rtp_pkt.size =
-		frame_size + RTP_TCP_HEAD_SIZE + RTP_HEADER_SIZE + AU_SIZE;
+	rtp_pkt.size = static_cast<uint16_t>(frame_size) + RTP_TCP_HEAD_SIZE +
+		       RTP_HEADER_SIZE + AU_SIZE;
 	rtp_pkt.last = 1;
 	uint8_t *rtp_pkt_data =
 		rtp_pkt.data.get() + RTP_TCP_HEAD_SIZE + RTP_HEADER_SIZE;
